# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APi для организаций",
        "version": "1.0.0"
    },
    "paths": {
        "/api/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Get list of buildings",
                "description": "Returns a list of buildings with their organizations",
                "operationId": "704d1f7897ed7a3b8d29dacea66823ad",
                "responses": {
                    "200": {
                        "description": "List of buildings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Building"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/organizations/building/{buildingId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations by building",
                "description": "Returns organizations based on building ID, including phones and activities",
                "operationId": "3af13fa794ae285a2e56257205c6f009",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of organizations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/organizations/activity/{activityId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations by activity",
                "description": "Returns organizations related to a specific activity, including phones",
                "operationId": "459e0ae0dadbde7ff24ad1607253485a",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of organizations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/organizations/search/name/{name}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Search organizations by name",
                "description": "Search for organizations by name and returns matching results",
                "operationId": "aa2243745ca5766d32487dd2c9e12a52",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of organizations matching the name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organization by ID",
                "description": "Returns detailed information about an organization by its ID",
                "operationId": "68e75650b204083886e125e02aaa2f4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organization not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/organizations/activity-tree/{activityId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations by activity tree",
                "description": "Returns organizations related to the activity and its children",
                "operationId": "c78dd329bc214279cf3ca5b7d7620b6d",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of organizations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Building": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Building Name"
                    },
                    "organizations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Organization"
                        }
                    }
                },
                "type": "object"
            },
            "Organization": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Organization Name"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "123-456-7890"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Buildings",
            "description": "Buildings"
        },
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}